refactor(web): extract Pydantic schemas to web/schemas/workflows.py (#167)
Move WorkflowJobInput / WorkflowCreateRequest / SweepSpecRequest /
WorkflowRunRequest out of the 1863-LOC workflows router into a new
web/schemas/workflows.py module so service and router modules can share
DTOs without a cycle. Router re-exports them for backward compatibility.
