+++ src/srunx/web/schemas/__init__.py
@@ -1,0 +1,19 @@
+"""Pydantic request/response schemas for the web API.
+
+Centralized here so service and router modules can share the same DTO
+definitions without router↔service import cycles.
+"""
+
+from .workflows import (
+    SweepSpecRequest,
+    WorkflowCreateRequest,
+    WorkflowJobInput,
+    WorkflowRunRequest,
+)
+
+__all__ = [
+    "SweepSpecRequest",
+    "WorkflowCreateRequest",
+    "WorkflowJobInput",
+    "WorkflowRunRequest",
+]
